项目三：product-lifecycle-prediction（1.1）/requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
torch
openpyxl
PyYAML
warnings